Download test reports and logs even when tests fail
Currently, the verifier plugin calls sys.exit() when robot returns a
non-zero exit code, which robot will do by default when at least one
test fails.  Since molecule exits, the failed test reports and logs are
not downloaded, and if the test instance is destroyed, the test logs are
lost.

Fix this by removing the exit() and always running the playbook to
download the logs and reports after running robot, even when it fails.
Add a error log after downloading the report to show the verification
failed.