Adding robotester support for Solaris11 (#1)
Some tweaks were needed.
Timestamp rollover test failed on Solaris11 with python2:
OverflowError: Python int too large to convert to C long

Setting explicitly pip executable to pip-3.5 (no pip3 link on Solaris)
with a variable fix this. Then python3 is used to install the needed
modules. This variable needs to be set to pip for centos-7.
For all other platforms the executable: parameter is set to pip3.