Improve support for delegated drivers
The ssh host options may not be set when using a delegated driver, so
avoid key errors when the ansible_ssh options are not present.  Also, be
sure to convert the port number to a string when using it to set the
command line options, since the command line arguments are required to
be strings.