Add robot debugfile support
Download the debug log file, in addition to the logs and report.

The debug file is optional, but instead of creating a dynamic list of
files to download, when not present, create a debug file with a message
to let the user how to enable debug logging.

Also update the example configuration in the docs to show the debugfile
option.