Move git installation tasks to a role
Move the tasks to ensure git is installed to a new bundled role which
includes the right task depending on the current package manager.

This avoids a bunch of "skipped" messages in the ansible output as well
avoids a duplication of the when clauses.  This also adds a a warning if
we don't recognize the ansible_pkg_mgr, instead of just silently
ignoring it.  We don't consider that situation an error, since developer
may have already installed git.