Add action_args argument for molecule 4 support
Molecule 4 added the optional action_args argument to the
execute() methods.  Add the new action_args to our verifier
execute() method to fix the exception thrown when running
with molecule 4.0.  For now, we do not use the action_args,
so we are compatible with older versions of molecule.

Fixes the TypeError exception when running with molecule 4.0.

        self._config.verifier.execute(action_args)
    TypeError: execute() takes 1 positional argument but 2 were given

Update the required molecule version to allow our package to be
installed with molecule 4.0.