New verifier options keyword
Change the 'options' keyword to be the plugin specific configuation and
rename the old 'options' keyword to 'robot' (since these are
specifically options for the robot command).  Move all of the plugin
keywords under options.  Update the code, the playbooks, the test
molecule files, and the docs.

This configuration layout change is needed since newer versions of
molecule are more strict on the schema validation. Fortunately the
schema specifies an 'options' keyword which is an object type, so we can
supply a dictionary with plugin specific keys.