Add patch_molecule_schema.py script
Unfortunately, molecule broke the schema validation for plugins in
molecule version 4.0.2.  As a workaround, provide a script to update the
schema file installed by molecule by adding the name of this plugin to
the hardcoded list of verifiers.

This script does some whitespace reformatting of the molecule.json file,
but that should be ok since this is just a work-around.