Use ssh when ansible_connection is smart
The default ansible_connnection is 'smart'.  Molecule sets the
ansible_connection to 'smart' when using the default driver (no driver,
and create/destroy playbooks are specified).  Assume it is ok to use ssh
to connect when the connection type is 'smart'.

Fixes the error message:

    CRITICAL Unsupported connection smart

when no driver is specified.