Add support for library uploads
Add support for uploading the libraries from the controller so local
libraries can be used without uploading to pypi or a private repo.

Treat the list of libraries as a list of dicts. When the dict has a
'name:' key use that to install the library with pip, as before. When
the item has a 'file:' key, use the value as the path to a local file to
be uploaded and installed.

For backward compat, if the item is not a dict, treat it as the package
name, as before.